requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
PyJWT==2.0.1
SQLAlchemy==1.3.22
Werkzeug==1.0.1